moved event list back
