+++ lib/backend/master_list.dart
@@ -1,5 +1,5 @@
 
-import 'package:executive_planner/backend/events/event_list.dart';
+import 'package:executive_planner/backend/event_list.dart';
 import 'package:executive_planner/backend/file_io.dart';
 import 'package:executive_planner/backend/jason.dart';
