+++ lib/pages/forms/event_add_form.dart
@@ -1,4 +1,4 @@
-import 'package:executive_planner/backend/events/event_list.dart';
+import 'package:executive_planner/backend/event_list.dart';
 import 'package:executive_planner/backend/misc.dart';
 import 'package:executive_planner/pages/forms/event_form.dart';
 import 'package:executive_planner/widgets/tag_selector.dart';
