+++ lib/widgets/bottom_nav_bar.dart
@@ -1,4 +1,4 @@
-import 'package:executive_planner/backend/events/event_list.dart';
+import 'package:executive_planner/backend/event_list.dart';
 import 'package:executive_planner/pages/calendar.dart';
 import 'package:flutter/material.dart';
