+++ lib/widgets/tag_selector.dart
@@ -1,5 +1,5 @@
 
-import 'package:executive_planner/backend/events/event_list.dart';
+import 'package:executive_planner/backend/event_list.dart';
 import 'package:executive_planner/backend/recurrence.dart';
 import 'package:executive_planner/backend/tag_model.dart';
 import 'package:executive_planner/widgets/search.dart';
